major update

+++ src/towerdefense/entity/tile/AbstractTile.java
@@ -3,7 +3,7 @@ package towerdefense.entity.tile;
 import towerdefense.entity.AbstractEntity;
 
 public abstract class AbstractTile extends AbstractEntity {
-    public AbstractTile(long createdTick, int x, int y, int width, int height) {
+    public AbstractTile(long createdTick, double x, double y, int width, int height) {
         super(createdTick, x, y, width, height);
     }
 }
